Refactor ControlStrip
